+++ views/modules/page/page_layout.blade.php
@@ -7,29 +7,32 @@
 @section('content')
     <div class="page-layout" data-id="{{ $page->id }}">
 
-        @include('core::contextualLinks')
-
         <?php 
         $inputDisabled = '';
         if(Auth::user()->cannot('manage layouts regions')) $inputDisabled = 'disabled';
         ?>
         <div class="region js-region skeleton d-none">
             <div class="inner">
-                <input type="text" {{ $inputDisabled }} name="">
+                <input type="text" class="name" {{ $inputDisabled }}>
+                <input type="hidden" class="width" {{ $inputDisabled }}>
+                <input type="hidden" class="height" {{ $inputDisabled }}>
                 @can('manage page layouts')
-                    <a href="{{ $deleteRegionUri }}" class="js-delete">Delete</a>
+                    <a href="{{ $deleteRegionUri }}" data-confirmmessage="This action cannot be undone" class="delete js-ajax-confirm-link" data-ajaxmethod="delete" data-confirmtitle="Delete region ?">Delete</a>
                 @endcan
             </div>
         </div>
 
+        {{ FormFacade::open(['url' => $saveRegionUri, 'method' => 'patch', 'class' => 'js-ajax-form']) }}
         <div id="page-container" class="page js-page">
             @if($regions)
                 @foreach($regions as $index => $region)
                     <div class="region js-region" style="width:calc({{ $region->width }}% - 20px);height:{{ $region->height }}px" data-width="{{ $region->width }}" data-height="{{ $region->height }}" data-id="{{ $region->id }}">
                         <div class="inner">
-                            <input type="text" {{ $inputDisabled }} name="regions[{{ $region->id }}][name]" value="{{ $region->name }}">
+                            <input type="text" class="name" {{ $inputDisabled }} name="models[{{ $region->id }}][name]" value="{{ $region->name }}">
+                            <input type="hidden" class="width" {{ $inputDisabled }} name="models[{{ $region->id }}][width]" value="{{ $region->width }}">
+                            <input type="hidden" class="height" {{ $inputDisabled }} name="models[{{ $region->id }}][height]" value="{{ $region->height }}">
                             @can('manage page layouts')
-                                <a href="{{ $region::transformUri('delete', [$region], config('core.ajaxPrefix')) }}" class="js-delete">Delete</a>
+                                <a href="{{ $region::makeUri('delete', [$region], adminPrefix()) }}" class="js-ajax-confirm-link delete" data-confirmtitle="Delete region ?" data-ajaxmethod="delete" data-confirmmessage="This action cannot be undone">Delete</a>
                             @endcan
                         </div>
                     </div>
@@ -39,10 +42,11 @@
 
         @can('manage page layouts')
             <div class="actions">
-                <a href="{{ $addRegionUri }}" class="js-add-region">Add a region</a>
-                <a href="{{ $saveRegionUri }}" class="js-save btn btn-primary float-right disabled">Save</a>
+                <a href="{{ $addRegionUri }}" class="add-region js-ajax-link-form">Add a region</a>
+                {{ FormFacade::submit('Save', ['class' => 'js-save btn btn-primary float-right']) }}
             </div>
         @endcan
+        {{ FormFacade::close() }}
 
     </div>
 @endsection
