+++ views/modules/taxonomy/edit-items.blade.php
@@ -7,35 +7,41 @@
 @section('content')
     <div class="edit-taxonomy-items" data-taxonomy="{{ $taxonomy->id }}">
 
-        @include('core::contextualLinks')
-
         <li class="list-group-item taxonomy-item skeleton d-none" data-item="">
           <div class="header">
             <i class="fa fa-bars"></i>
             <span class="name"></span>
+            <input type="hidden" name="models[id][weight]">
             @can('delete taxonomy terms')
-              <a href="{{ $deleteItemUri }}" class="js-delete float-right">Delete</a>
+              <a href="{{ $deleteItemUri }}" class="js-delete js-ajax-confirm-link float-right" data-ajaxmethod="delete" data-confirmmessage="This action cannot be undone">Delete</a>
             @endcan
             @can('edit taxonomy terms')
-              <a href="{{ $editItemUri }}" class="js-edit float-right mr-1">Edit</a>
+              <a href="{{ $editItemUri }}" class="js-edit js-ajax-link-form float-right mr-1">Edit</a>
             @endcan
           </div>
         </li>
 
         @can('add taxonomy terms')
-          <a href="{{ $addItemUri }}" class="js-add mb-2 d-inline-block">Add an item</a>
+          <a href="{{ $addItemUri }}" class="js-add mb-2 d-inline-block js-ajax-link-form">Add an item</a>
+        @endcan
+
+        @can('edit taxonomy terms')
+          {{ FormFacade::open(['url' => $patchItemsUri, 'method' => 'patch', 'class' => 'form js-ajax-form js-show-message']) }}
         @endcan
 
         <div class="taxonomy-tree">
           @if(!$items->isEmpty())
-            @include('taxonomy::tree')
+            <ul class="list-group taxonomy-list">
+              @include('taxonomy::tree')
+            </ul>
           @else
             <p>No items</p>
           @endif
        	</div>
 
         @can('edit taxonomy terms')
-          <a href="{{ $patchItemsUri }}" class="js-save mt-2 d-inline-block btn btn-primary float-right disabled">Save</a>
+          {{ FormFacade::submit('Save', ['class' => 'save mt-2 d-inline-block btn btn-primary float-right']) }}
+          {{ FormFacade::close() }}
         @endcan
 
     </div>
